script changes

+++ demo-10/scripts/volumes.sh
@@ -3,11 +3,12 @@
 vgchange -ay
 
 DEVICE_FS=`blkid -o value -s TYPE ${DEVICE}`
-exit 0
-pvcreate ${DEVICE}
-vgcreate data ${DEVICE}
-lvcreate --name volume1 -l 100%FREE data
-mkfs.ext4 /dev/data/volume1
+if [ "`echo -n $DEVICE_FS`" == "" ] ; then 
+	pvcreate ${DEVICE}
+	vgcreate data ${DEVICE}
+	lvcreate --name volume1 -l 100%FREE data
+	mkfs.ext4 /dev/data/volume1
+fi
 mkdir -p /data
-echo '/dev/data/volume /data defaults 0 0' >> /etc/fstab
+echo '/dev/data/volume1 /data ext4 defaults 0 0' >> /etc/fstab
 mount /data
